refactor(craft): extract Craft persistence into CraftLibraryRepository (refs #601)
Move the five Craft audio lifecycle methods (dedupe lookup, import,
load-for-edit, in-place rewrite, history-record removal) out of
MediaLibraryRepository into a deep craft-feature module. The dedupe
content hash had three copies; it is now one private helper shared by
import / find / update. CraftEditSource moves into the craft domain.

Also extract the ffmpeg duration probe into a shared
probeAndPatchMediaDuration helper (two real callers: library import and
craft import/update) and delete the no-op ensureVideoPosterAfterMetadataInsert
hook plus CloudAddToLibrary's now-unneeded repository dependency.
