+++ test/features/craft/presentation/craft_screen_test.dart
@@ -18,8 +18,8 @@ import 'package:enjoy_player/features/craft/domain/translation_style.dart';
 import 'package:enjoy_player/features/craft/presentation/advanced_tools.dart';
 import 'package:enjoy_player/features/craft/presentation/craft_screen.dart';
 import 'package:enjoy_player/features/craft/presentation/express_flow.dart';
-import 'package:enjoy_player/features/library/application/library_repository_provider.dart';
-import 'package:enjoy_player/features/library/data/library_repository.dart';
+import 'package:enjoy_player/features/craft/application/craft_library_repository_provider.dart';
+import 'package:enjoy_player/features/craft/data/craft_library_repository.dart';
 import 'package:enjoy_player/l10n/app_localizations.dart';
 
 // === Fakes ===
@@ -72,7 +72,7 @@ class _FakeTranscriber implements CraftTranscriber {
   }) async => 'I had a great day today.';
 }
 
-class _FakeLibraryRepository implements MediaLibraryRepository {
+class _FakeLibraryRepository implements CraftLibraryRepository {
   @override
   dynamic noSuchMethod(Invocation invocation) => null;
 }
@@ -97,7 +97,7 @@ List<Override> _baseOverrides() => [
   craftTranslatorProvider.overrideWithValue(_FakeTranslator()),
   craftSynthesizerProvider.overrideWithValue(_FakeSynthesizer()),
   craftTranscriberProvider.overrideWithValue(_FakeTranscriber()),
-  mediaLibraryRepositoryProvider.overrideWithValue(_FakeLibraryRepository()),
+  craftLibraryRepositoryProvider.overrideWithValue(_FakeLibraryRepository()),
 ];
 
 void main() {
