+++ test/features/library/library_repository_extra_test.dart
@@ -1,6 +1,5 @@
 import 'dart:convert';
 import 'dart:io';
-import 'dart:typed_data';
 
 import 'package:crypto/crypto.dart';
 import 'package:cross_file/cross_file.dart';
@@ -1208,203 +1207,6 @@ void main() {
 
         await ytRepo.refreshYoutubeMetadataIfNeeded(mediaId);
         expect(syncCalls, 0);
-      });
-    });
-
-    group('importCraftedFromText additional', () {
-      test('uses provided primaryTimelineJson', () async {
-        final timeline = jsonEncode([
-          {'text': 'Hello', 'start': 0, 'duration': 500},
-          {'text': 'world', 'start': 500, 'duration': 400},
-        ]);
-
-        final id = await repo.importCraftedFromText(
-          audioBytes: Uint8List.fromList([1, 2, 3]),
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Hello world',
-          normalizedText: 'Hello world',
-          primaryTimelineJson: timeline,
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final transcripts = await db.transcriptDao.listForTarget('Audio', id);
-        expect(transcripts, hasLength(1));
-        expect(transcripts.first.timelineJson, timeline);
-      });
-
-      test('omits transcript when primaryTimelineJson is null', () async {
-        final id = await repo.importCraftedFromText(
-          audioBytes: Uint8List.fromList([4, 5, 6]),
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Blank timeline',
-          normalizedText: 'Blank timeline',
-          primaryTimelineJson: null,
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final transcripts = await db.transcriptDao.listForTarget('Audio', id);
-        expect(transcripts, isEmpty);
-        final audio = await db.audioDao.getById(id);
-        expect(audio, isNotNull);
-      });
-
-      test('voice participates in dedupe key', () async {
-        final audioBytes = Uint8List.fromList([7, 8, 9]);
-
-        final id1 = await repo.importCraftedFromText(
-          audioBytes: audioBytes,
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Voice test',
-          normalizedText: 'Voice test',
-          voice: 'alloy',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final id2 = await repo.importCraftedFromText(
-          audioBytes: audioBytes,
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Voice test',
-          normalizedText: 'Voice test',
-          voice: 'nova',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        expect(id1, isNot(id2));
-      });
-
-      test('same voice dedupes correctly', () async {
-        final audioBytes = Uint8List.fromList([10, 11]);
-
-        final id1 = await repo.importCraftedFromText(
-          audioBytes: audioBytes,
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Same voice',
-          normalizedText: 'Same voice',
-          voice: 'alloy',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final id2 = await repo.importCraftedFromText(
-          audioBytes: audioBytes,
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Same voice',
-          normalizedText: 'Same voice',
-          voice: 'alloy',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        expect(id1, id2);
-      });
-
-      test('enqueues sync create', () async {
-        final syncLog = <(SyncEntityType, String, SyncAction)>[];
-        final syncRepo = MediaLibraryRepository(
-          db,
-          FileStorage(),
-          enqueueSync: (type, id, action) async {
-            syncLog.add((type, id, action));
-          },
-        );
-
-        final id = await syncRepo.importCraftedFromText(
-          audioBytes: Uint8List.fromList([20]),
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Sync craft',
-          normalizedText: 'Sync craft',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        expect(
-          syncLog,
-          contains((SyncEntityType.audio, id, SyncAction.create)),
-        );
-      });
-
-      test('title is not truncated when text is short', () async {
-        final shortText = 'Short';
-        final id = await repo.importCraftedFromText(
-          audioBytes: Uint8List.fromList([30]),
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: shortText,
-          normalizedText: shortText,
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final row = await db.audioDao.getById(id);
-        expect(row!.title, shortText);
-      });
-    });
-
-    group('findExistingCrafted', () {
-      test('voice differentiates lookup', () async {
-        await repo.importCraftedFromText(
-          audioBytes: Uint8List.fromList([40]),
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Voice find',
-          normalizedText: 'Voice find',
-          voice: 'alloy',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final withSameVoice = await repo.findExistingCrafted(
-          learningLanguage: 'en',
-          normalizedText: 'Voice find',
-          sourceFlag: 'craft-direct',
-          voice: 'alloy',
-        );
-        expect(withSameVoice, isNotNull);
-
-        final withDifferentVoice = await repo.findExistingCrafted(
-          learningLanguage: 'en',
-          normalizedText: 'Voice find',
-          sourceFlag: 'craft-direct',
-          voice: 'nova',
-        );
-        expect(withDifferentVoice, isNull);
-      });
-
-      test('sourceFlag differentiates lookup', () async {
-        await repo.importCraftedFromText(
-          audioBytes: Uint8List.fromList([50]),
-          audioFormat: 'wav',
-          learningLanguage: 'en',
-          text: 'Flag test',
-          normalizedText: 'Flag test',
-          sourceFlag: 'craft-direct',
-          signedInUserId: _testUserId,
-        );
-
-        final sameFlag = await repo.findExistingCrafted(
-          learningLanguage: 'en',
-          normalizedText: 'Flag test',
-          sourceFlag: 'craft-direct',
-        );
-        expect(sameFlag, isNotNull);
-
-        final differentFlag = await repo.findExistingCrafted(
-          learningLanguage: 'en',
-          normalizedText: 'Flag test',
-          sourceFlag: 'craft-translate',
-        );
-        expect(differentFlag, isNull);
       });
     });
